+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- `data-*` attribute to `Menu` component (`data-state="open"`, `data-state="active"`, etc).
+
 ## [0.2.0] - 2023-04-30
 
 ### Added
@@ -21,4 +26,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed `Listbox` component.
 - Removed `Tabs` component.
 
+[unreleased]: https://github.com/abeidahmed/ember-primitiveui/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/abeidahmed/ember-primitiveui/releases/tag/v0.2.0